Move respond.go into main render.go
